Replace nginx with HAProxy

+++ contrib/docker-compose/standalone/init-letsencrypt.bash
@@ -26,22 +26,28 @@
 
 HERE=$(cd "$(dirname "$0")"; pwd)
 set -eu
-[ -e "$HERE/.env" ]
-. "$HERE/.env"
+[ -f "$HERE/.env" ] || { echo "Environment file not found; you must run ./mksite.bash first"; exit 1; }
 
 if ! [ -x "$(command -v docker-compose)" ]; then
   echo 'Error: docker-compose is not installed.' >&2
   exit 1
 fi
 
+FQDN=$(awk -F= '/^FQDN=/ {print $2}' < "$HERE/.env" | tail -1)
+ALIAS_FQDNS=$(awk -F= '/^ALIAS_FQDNS=/ {print $2; exit}' < "$HERE/.env" | tail -1)
+
+# Strip enclosing quotes, as docker-compose<1.29 does not parse shell metachars in .env
+# See https://github.com/docker/compose/issues/8388
+ALIAS_FQDNS="${ALIAS_FQDNS%\"}"
+ALIAS_FQDNS="${ALIAS_FQDNS#\"}"
+
 domains=($FQDN $ALIAS_FQDNS)
 rsa_key_size=4096
-email="$EMAIL" # Adding a valid address is strongly recommended
+email=$(awk -F= '/^EMAIL=/ {print $2}' < "$HERE/.env" | tail -1)
 
 echo "### Downloading recommended TLS parameters ..."
 docker-compose run --rm --entrypoint "/bin/sh -c \"\
   cd /etc/letsencrypt && \
-  wget -q https://raw.githubusercontent.com/certbot/certbot/master/certbot-nginx/certbot_nginx/_internal/tls_configs/options-ssl-nginx.conf -O options-ssl-nginx.conf && \
   wget -q https://ssl-config.mozilla.org/ffdhe2048.txt -O ssl-dhparams.pem\"" certbot
 echo
 
@@ -57,8 +63,8 @@ for domain in "${domains[@]}"; do
   echo
 done
 
-echo "### Starting nginx ..."
-docker-compose up --force-recreate -d nginx
+echo "### Shut down deployment ..."
+docker-compose down
 echo
 
 echo "### Deleting dummy certificates for ${domains[*]} ..."
@@ -83,10 +89,10 @@ esac
 if [ "${CERTBOT_STAGING:-0}" != "0" ]; then staging_arg="--staging"; else staging_arg=""; fi
 
 # Use a non-default CA server if specified
-if [ -n "${ACME_SERVER}" ]; then server_arg="--server ${ACME_SERVER}"; else server_arg=""; fi
+if [ -n "${ACME_SERVER:-}" ]; then server_arg="--server ${ACME_SERVER}"; else server_arg=""; fi
 
-docker-compose run --rm --entrypoint "\
-  certbot certonly --webroot -w /etc/nginx/html \
+docker-compose run --rm -p 80:80 --entrypoint "\
+  certbot certonly --standalone \
     $server_arg \
     $staging_arg \
     $email_arg \
@@ -96,5 +102,4 @@ docker-compose run --rm --entrypoint "\
     --force-renewal" certbot
 echo
 
-echo "### Shutting down ..."
-docker-compose down
+echo "### Done"
